EES-2880 Removing ImporterUtils by creating SoloImportableLevels to avoid naming confusion between methods.
